Organize postprocess module into subdirectories
Move processors into category subdirectories for better organization:
- common/: General-purpose processors (strip_links, strip_colour)
- wandering_inn/: Series-specific processors (mrsha_write)

This allows adding new processor categories by simply creating new
subdirectories with their own mod.rs files.
