+++ src/postprocess/common/mod.rs
@@ -1,0 +1,5 @@
+mod strip_colour;
+mod strip_links;
+
+pub use strip_colour::StripColourProcessor;
+pub use strip_links::StripLinksProcessor;
